# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Alset + LevelUpAuto unified API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alset/vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List vehicles for current owner
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Register a Tesla vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleBody"
      responses:
        "201":
          description: Vehicle registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /alset/vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"

  # ─── Alset Claims ──────────────────────────────────────────────────────────
...
  /alset/work-orders:
    get:
      operationId: listWorkOrders
      tags: [work-orders]
      summary: List work orders
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [work-orders]
      summary: Create a new work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkOrderBody"
      responses:
        "201":
          description: Work order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /alset/work-orders/{id}:
    get:
      operationId: getWorkOrder
      tags: [work-orders]
      summary: Get a work order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    patch:
      operationId: updateWorkOrder
      tags: [work-orders]
      summary: Update work order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkOrderBody"
      responses:
        "200":
          description: Work order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"

  # ─── Alset Towing ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        vin:
          type: string
        model:
          type: string
          enum: [Model S, Model 3, Model X, Model Y, Cybertruck, Roadster]
        year:
          type: integer
        color:
          type: string
        licensePlate:
          type: string
        mileage:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, vin, model, year, color, licensePlate, createdAt]
    CreateVehicleBody:
      type: object
      properties:
        vin:
          type: string
        model:
          type: string
          enum: [Model S, Model 3, Model X, Model Y, Cybertruck, Roadster]
        year:
          type: integer
        color:
          type: string
        licensePlate:
          type: string
        mileage:
          type: integer
          nullable: true
      required: [vin, model, year, color, licensePlate]
...
    WorkOrder:
      type: object
      properties:
        id:
          type: integer
        workOrderNumber:
          type: string
        claimId:
          type: integer
          nullable: true
        vehicleId:
          type: integer
        vehicleVin:
          type: string
          nullable: true
        vehicleModel:
          type: string
          nullable: true
        shopId:
          type: integer
          nullable: true
        shopName:
          type: string
          nullable: true
        technicianName:
          type: string
          nullable: true
        description:
          type: string
        laborHours:
          type: number
          nullable: true
        partsTotal:
          type: number
          nullable: true
        laborRate:
          type: number
          nullable: true
        totalCost:
          type: number
          nullable: true
        status:
          type: string
          enum: [draft, assigned, in-progress, awaiting-parts, completed, invoiced]
        startDate:
          type: string
          format: date
          nullable: true
        completionDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, workOrderNumber, vehicleId, description, status, createdAt, updatedAt]
    CreateWorkOrderBody:
      type: object
      properties:
        claimId:
          type: integer
          nullable: true
        vehicleId:
          type: integer
        description:
          type: string
        laborHours:
          type: number
          nullable: true
        partsTotal:
          type: number
          nullable: true
        laborRate:
          type: number
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required: [vehicleId, description]
    UpdateWorkOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, assigned, in-progress, awaiting-parts, completed, invoiced]
        technicianName:
          type: string
          nullable: true
        laborHours:
          type: number
          nullable: true
        partsTotal:
          type: number
          nullable: true
        completionDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
...