# lib/api-zod/src/generated/types/createVehicleBodyModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alset + LevelUpAuto unified API specification
 * OpenAPI spec version: 0.2.0
 */

export type CreateVehicleBodyModel =
  (typeof CreateVehicleBodyModel)[keyof typeof CreateVehicleBodyModel];

export const CreateVehicleBodyModel = {
  Model_S: "Model S",
  Model_3: "Model 3",
  Model_X: "Model X",
  Model_Y: "Model Y",
  Cybertruck: "Cybertruck",
  Roadster: "Roadster",
} as const;