# lib/api-zod/src/generated/types/createWorkOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alset + LevelUpAuto unified API specification
 * OpenAPI spec version: 0.2.0
 */

export interface CreateWorkOrderBody {
  claimId?: number | null;
  vehicleId: number;
  description: string;
  laborHours?: number | null;
  partsTotal?: number | null;
  laborRate?: number | null;
  startDate?: Date | null;
  notes?: string | null;
}