# lib/api-zod/src/generated/types/updateWorkOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alset + LevelUpAuto unified API specification
 * OpenAPI spec version: 0.2.0
 */
import type { UpdateWorkOrderBodyStatus } from "./updateWorkOrderBodyStatus";

export interface UpdateWorkOrderBody {
  status?: UpdateWorkOrderBodyStatus;
  technicianName?: string | null;
  laborHours?: number | null;
  partsTotal?: number | null;
  completionDate?: Date | null;
  notes?: string | null;
}